Organize search utils and update references
