+++ frontend/tests/e2e/search/search-core/specs/search-copy-url.spec.ts
@@ -7,8 +7,8 @@
 import { expect, test } from "@playwright/test";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { waitForURLContainsQueryParam } from "tests/e2e/playwrightUtils";
-import { fillSearchInputAndSubmit } from "tests/e2e/search/searchSpecUtil";
 import { VALID_TAGS } from "tests/e2e/tags";
+import { fillSearchInputAndSubmit } from "tests/e2e/utils/search/searchSpecUtil";
 
 const { baseUrl } = playwrightEnv;
