+++ frontend/tests/e2e/search/search-core/specs/search-default-filters.spec.ts
@@ -5,8 +5,8 @@
  */
 
 import { expect, test } from "@playwright/test";
-import { expectCheckboxIDIsChecked } from "tests/e2e/search/searchSpecUtil";
 import { VALID_TAGS } from "tests/e2e/tags";
+import { expectCheckboxIDIsChecked } from "tests/e2e/utils/search/searchSpecUtil";
 
 const { GRANTEE, OPPORTUNITY_SEARCH, SMOKE } = VALID_TAGS;
