+++ frontend/tests/e2e/search/search-core/specs/search-results.spec.ts
@@ -6,11 +6,11 @@
 
 import { expect, test } from "@playwright/test";
 import { waitForURLContainsQueryParamValue } from "tests/e2e/playwrightUtils";
+import { VALID_TAGS } from "tests/e2e/tags";
 import {
   fillSearchInputAndSubmit,
   waitForSearchResultsInitialLoad,
-} from "tests/e2e/search/searchSpecUtil";
-import { VALID_TAGS } from "tests/e2e/tags";
+} from "tests/e2e/utils/search/searchSpecUtil";
 
 const { GRANTEE, OPPORTUNITY_SEARCH, SMOKE, CORE_REGRESSION } = VALID_TAGS;
