+++ frontend/tests/e2e/search/search-state/specs/search-state-persistence.spec.ts
@@ -18,6 +18,7 @@ import {
   waitForURLContainsQueryParamValue,
   waitForURLContainsQueryParamValues,
 } from "tests/e2e/playwrightUtils";
+import { VALID_TAGS } from "tests/e2e/tags";
 import {
   ensureAccordionExpanded,
   ensureFilterDrawerOpen,
@@ -32,8 +33,7 @@ import {
   toggleFilterDrawer,
   waitForFilterOptions,
   waitForSearchResultsInitialLoad,
-} from "tests/e2e/search/searchSpecUtil";
-import { VALID_TAGS } from "tests/e2e/tags";
+} from "tests/e2e/utils/search/searchSpecUtil";
 
 const { GRANTEE, OPPORTUNITY_SEARCH, FULL_REGRESSION, CORE_REGRESSION } =
   VALID_TAGS;
